added surface generation

+++ Minecraft/Blocks/Block.cs
@@ -7,9 +7,9 @@ namespace Minecraft
     public class Block
     {
         public BlockType Type { get; protected set; }
-        public int XPOS { get; set; }
-        public int YPOS { get; set; }
-        public int ZPOS { get; set; }
+        public float XPOS { get; set; }
+        public float YPOS { get; set; }
+        public float ZPOS { get; set; }
 
     }
 }
